Removed encrypt and prove, added randomise and prove
