+++ src/public.rs
@@ -1,0 +1,213 @@
+use clear_on_drop::clear::Clear;
+use curve25519_dalek::constants::{RISTRETTO_BASEPOINT_POINT, RISTRETTO_BASEPOINT_COMPRESSED};
+use curve25519_dalek::ristretto::{CompressedRistretto, RistrettoPoint};
+use curve25519_dalek::scalar::Scalar;
+use rand_core::{OsRng, };
+use serde::{Deserialize, Serialize};
+use sha2::{Digest, Sha512};
+
+use zkp::{Transcript, CompactProof, };
+
+use crate::ciphertext::*;
+
+/// The `PublicKey` struct represents an ElGamal public key.
+#[derive(Serialize, Deserialize, Copy, Clone, Debug)]
+pub struct PublicKey(RistrettoPoint);
+
+define_proof! {dl_knowledge, "DLKnowledge Proof", (x), (A), (G) : A = (x * G)}
+define_proof! {dleq, "DLEQ Proof", (x), (A, B, H), (G) : A = (x * B), H = (x * G)}
+
+impl PublicKey {
+    /// Encrypts a message in the Ristretto group. It has the additive homomorphic property,
+    /// allowing addition (and subtraction) by another ciphertext and multiplication (and division)
+    /// by scalars.
+    ///
+    /// #Example
+    /// ```
+    /// extern crate rand;
+    /// extern crate curve25519_dalek;
+    /// extern crate elgamal_ristretto;
+    /// use rand_core::OsRng;
+    /// use elgamal_ristretto::public::{PublicKey, };
+    /// use elgamal_ristretto::private::{SecretKey, };
+    /// use curve25519_dalek::ristretto::{RistrettoPoint, };
+    /// use curve25519_dalek::scalar::{Scalar, };
+    ///
+    /// # fn main() {
+    ///        let mut csprng = OsRng;
+    ///        // Generate key pair
+    ///        let sk = SecretKey::new(&mut csprng);
+    ///        let pk = PublicKey::from(&sk);
+    ///
+    ///        // Generate random messages
+    ///        let ptxt1 = RistrettoPoint::random(&mut csprng);
+    ///        let ptxt2 = RistrettoPoint::random(&mut csprng);
+    ///
+    ///        // Encrypt messages
+    ///        let ctxt1 = pk.encrypt(ptxt1);
+    ///        let ctxt2 = pk.encrypt(ptxt2);
+    ///
+    ///        // Add ciphertexts and check that addition is maintained in the plaintexts
+    ///        let encrypted_addition = ctxt1 + ctxt2;
+    ///        let decrypted_addition = sk.decrypt(encrypted_addition);
+    ///
+    ///        assert_eq!(ptxt1 + ptxt2, decrypted_addition);
+    ///
+    ///        // Multiply by scalar and check that multiplication is maintained in the plaintext
+    ///        let scalar_mult = Scalar::random(&mut csprng);
+    ///        assert_eq!(sk.decrypt(ctxt1 * scalar_mult), scalar_mult * ptxt1);
+    /// # }
+    /// ```
+    pub fn encrypt(self, message: RistrettoPoint) -> Ciphertext {
+        let mut csprng: OsRng = OsRng;
+        let mut random: Scalar = Scalar::random(&mut csprng);
+
+        let random_generator = &RISTRETTO_BASEPOINT_POINT * &random;
+        let encrypted_plaintext = message + &self.0 * &random;
+        random.clear();
+        Ciphertext {
+            pk: self,
+            points: (random_generator, encrypted_plaintext),
+        }
+    }
+
+    /// Get the public key as a RistrettoPoint
+    pub fn get_point(&self) -> RistrettoPoint {
+        self.0
+    }
+
+    /// Verify EdDSA signature
+    ///
+    /// #Example
+    /// ```
+    /// extern crate rand;
+    /// extern crate curve25519_dalek;
+    /// extern crate elgamal_ristretto;
+    /// use rand_core::OsRng;
+    /// use elgamal_ristretto::public::{PublicKey, };
+    /// use elgamal_ristretto::private::{SecretKey, };
+    /// use curve25519_dalek::ristretto::RistrettoPoint;
+    ///
+    /// # fn main() {
+    ///       // Generate key-pair
+    ///       let mut csprng = OsRng;
+    ///       let sk = SecretKey::new(&mut csprng);
+    ///       let pk = PublicKey::from(&sk);
+    ///
+    ///       // Sign message
+    ///       let msg = RistrettoPoint::random(&mut csprng);
+    ///       let signature = sk.sign(msg);
+    ///       // Verify signature
+    ///       assert!(pk.verify_signature(&msg, signature));
+    ///
+    ///       // Verify signature against incorrect message
+    ///       assert!(!pk.verify_signature(&RistrettoPoint::random(&mut csprng), signature))
+    /// # }
+    /// ```
+    pub fn verify_signature(self, message: &RistrettoPoint, signature: (Scalar, RistrettoPoint)) -> bool {
+        let verification_hash = Scalar::from_hash(
+            Sha512::new()
+                .chain(signature.1.compress().to_bytes())
+                .chain(self.0.compress().to_bytes())
+                .chain(message.compress().to_bytes()),
+        );
+
+        let check =
+            &signature.0 * &RISTRETTO_BASEPOINT_POINT == signature.1 + verification_hash * self.0;
+        check
+    }
+
+    /// Verify proof of knowledege of private key related to a public key
+    ///
+    /// Example
+    /// ```
+    /// extern crate rand;
+    /// extern crate curve25519_dalek;
+    /// extern crate elgamal_ristretto;
+    /// use rand_core::OsRng;
+    /// use elgamal_ristretto::public::{PublicKey, };
+    /// use elgamal_ristretto::private::{SecretKey, };
+    /// use curve25519_dalek::ristretto::RistrettoPoint;
+    ///
+    /// # fn main() {
+    ///       let mut csprng = OsRng;
+    ///       let sk = SecretKey::new(&mut csprng);
+    ///       let pk = PublicKey::from(&sk);
+    ///
+    ///       let proof = sk.prove_knowledge();
+    ///       assert!(pk.verify_proof_knowledge(proof));
+    /// # }
+    /// ```
+    pub fn verify_proof_knowledge(self, proof: CompactProof) -> bool {
+        let mut transcript = Transcript::new(b"ProveKnowledgeSK");
+        dl_knowledge::verify_compact(
+            &proof,
+            &mut transcript,
+            dl_knowledge::VerifyAssignments {
+                A: &self.0.compress(),
+                G: &RISTRETTO_BASEPOINT_COMPRESSED,
+            },
+        ).is_ok()
+    }
+
+    /// Verify correct decryption
+    ///
+    /// Example
+    /// ```
+    /// extern crate rand;
+    /// extern crate curve25519_dalek;
+    /// extern crate elgamal_ristretto;
+    /// use rand_core::OsRng;
+    /// use elgamal_ristretto::public::{PublicKey, };
+    /// use elgamal_ristretto::private::{SecretKey, };
+    /// use curve25519_dalek::ristretto::RistrettoPoint;
+    ///
+    /// # fn main() {
+    ///    let mut csprng = OsRng;
+    ///    let sk = SecretKey::new(&mut csprng);
+    ///    let pk = PublicKey::from(&sk);
+    ///
+    ///    let plaintext = RistrettoPoint::random(&mut csprng);
+    ///    let ciphertext = pk.encrypt(plaintext);
+    ///
+    ///    let decryption = sk.decrypt(ciphertext);
+    ///    let proof = sk.prove_correct_decryption(ciphertext, decryption);
+    ///
+    ///    assert!(pk.verify_correct_decryption(proof, ciphertext, decryption));
+    /// # }
+    /// ```
+    pub fn verify_correct_decryption(self, proof: CompactProof, ciphertext: Ciphertext, plaintext: RistrettoPoint) -> bool {
+        let mut transcript = Transcript::new(b"ProveCorrectDecryption");
+        dl_knowledge::verify_compact(
+            &proof,
+            &mut transcript,
+            dl_knowledge::VerifyAssignments {
+                A: &(ciphertext.points.1 - plaintext).compress(),
+                G: &ciphertext.points.0.compress(),
+            },
+        ).is_ok()
+    }
+
+    /// Convert to bytes
+    pub fn to_bytes(&self) -> [u8; 32] {
+        self.0.compress().to_bytes()
+    }
+
+    /// Generate public key from bytes
+    pub fn from_bytes(bytes: &[u8]) -> PublicKey {
+        PublicKey(CompressedRistretto::from_slice(bytes).decompress().unwrap())
+    }
+}
+
+impl From<RistrettoPoint> for PublicKey {
+    /// Given a secret key, compute its corresponding Public key
+    fn from(point: RistrettoPoint) -> PublicKey {
+        PublicKey(point)
+    }
+}
+
+impl PartialEq for PublicKey {
+    fn eq(&self, other: &PublicKey) -> bool {
+        self.0 == other.0
+    }
+}
